Python Bootscamp
The aim of this exercise is to evaluate my abilities for applying theoretical knowledge to a practical case in the universe of online advertising.

Read and clean the dataset (remove missing values, select relevant columns, and convert column types if necessary).
Explore data, create visualizations, and look for patterns. Try to find the relationship between the features and the occurrence of a click.
Use a statistical method that can identify the factors (features, columns) which seem to be the most influential for a click occurrence.