Chiffre de Vigenère
The Vigenère cipher is an encryption algorithm that uses a series of Caesar ciphers to form a cipher alphabet, and is a simple form of multi-table cipher.

Known for its simplicity and ease of use, the Vigenère cipher is often difficult for beginners to break, so it is also known as the "unbreakable cipher". This also makes many people use the Vigenère cipher to encrypt it in order to crack it.

This file consists of two parts, encryption and decryption.